Player carousel refactored and animation styles applied
