+++ src/components/atoms/arrows/arrows.js
@@ -1,0 +1,12 @@
+import styles from "./arrows.module.css";
+
+const Arrows = () => {
+
+    return (
+        <div className={styles.arrows}>
+        <div className={styles.arrowsContent}> *ARROWS HERE* </div>
+      </div>
+    );
+};
+
+export { Arrows };
