+++ src/components/atoms/player-card/player-card.js
@@ -3,9 +3,10 @@ import styles from './player-card.module.css';
 
 const PlayerCard = ({ name, position, img, handleClick }) => {
     return (
-      <div className={styles.card} 
-          onClick={() => handleClick(name)}
-        >
+      <main className={styles.card}
+      onClick={() => handleClick(name)}
+      >
+          <div className={styles.contents}>
             <Image 
               src={img}
               alt=""
@@ -15,7 +16,8 @@ const PlayerCard = ({ name, position, img, handleClick }) => {
                     <h1>{` ${name}`}</h1>
                     <p>{` ${position}`}</p>
                 </div>
-        </div>
+          </div>
+        </main>
     );
 };
