+++ src/components/atoms/player-card/player-card.module.css
@@ -1,3 +1,10 @@
+.contents {
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  width: 210px;
+  height: 100%;  
+}
 
 .card {
   display: flex;
@@ -10,6 +17,8 @@
   justify-content: space-evenly;
   margin: 0 2% 0;
   box-shadow: 0 4px 23px 5px;
+  gap: 10px;
+  z-index: 5;
 }
 
 .card:hover {
@@ -19,11 +28,11 @@
 
 @keyframes slidein {
     from {
-        scale: 100% ;
+        transform: scale(1) ;
       }
     
       to {
-        scale: 110% ;
+        transform: scale(1.1) ;
       }
 }
 
@@ -31,27 +40,35 @@
     margin: 0 5% 0;
 }
 
-.img {
-    justify-self: flex-end; 
-    border-radius: 8%;
-    border: 0.5rem thick ridge darkred;
-}
-
 .infosection {
     display: flex; 
     flex-direction: column;
-    align-items: flex-start;
-    justify-content: flex-end;
-    width: 80%; 
-    font-size: 100%;
+    padding: 0 5%;
     font-family:'Gill Sans', 'Gill Sans MT', Calibri, 'Trebuchet MS', sans-serif;
+    width: 100%;
+    height: 45%;
+    gap: 10px;
+    position: relative;
+    box-sizing: border-box;
+    align-items: center;
 }
 
 .infosection > h1 { 
-  font-size: 150%;
+  font-size: 130%;
   font-weight: normal;
 }
 
 .infosection > p { 
-  font-weight: normal;
+  font-weight: 650;
+  font-size: 90%;
+  position: absolute;
+  top: 55px;
+}
+
+.img {
+    justify-self: flex-end; 
+    border-radius: 8%;
+    height: 60%;
+    width: 100%;
+    padding: 10px 0;
 }
