+++ src/components/molecules/card-section/card-section.js
@@ -1,0 +1,18 @@
+import { Arrows } from "@/components/atoms/arrows/arrows";
+import { PlayerCarousel } from "../player-carousel/player-carousel";
+import styles from "./card-section.module.css";
+
+const Cardsection = () => {
+
+  return (
+    <div className={styles.wrapper}>
+    <div className={styles.gradientTop}>
+      <PlayerCarousel/>
+      <div className={styles.gradientBottom}></div>
+      <Arrows />
+    </div>
+    </div>
+  );
+};
+
+export { Cardsection };
