--- src/components/molecules/carousel/carousel.module.css
@@ -1,107 +1,0 @@
-.wrapper {
-    width: 100%;
-    height: 95vh;
-    background: linear-gradient(to top, rgba(235, 9, 9, 0), rgb(63, 6, 185));
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-end;
-    align-items: center;
-
-}
-
-.inner {
-    display: flex;
-    flex-direction: column;
-    width: 100%;
-    height: 100%;
-    border-radius: 10px;
-}
-
-.gradientBottom {
-    border-radius: 10px;
-    position: absolute;
-    top: 0;
-    right: 0; 
-    display: flex;
-    align-items: center;
-    width: 11%;
-    height: 100%;
-    background: linear-gradient(to right, rgb(255, 255, 255, 0), rgb(255, 255, 255)); 
-    opacity: 100%;
-}
-
-.gradientTop {
-    position: relative;
-    width: 100%;
-    height: 100%;
-    justify-content: center;
-    display: flex;
-    align-items: flex-end;
-}
-
-.cardRow {
-    display: flex;
-    height: 100%;
-    overflow: scroll;
-    overflow-x: scroll;
-    scrollbar-width: none;
-    align-items: center;
-    scroll-snap-type: x mandatory;
-}
-
-.arrows {
-    flex-direction: column;
-    border: solid 2mm black;
-    border-radius: 10%;
-    width: 20%;
-    height: 10%; 
-    align-self: center;
-    margin: 1rem 0 0 0;
-}
-
-.arrowsContent {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 100%;
-}
-
-.hologram {
-    opacity: 100;
-    height: 50%;
-    width: 100%;
-    background: linear-gradient(to top, rgba(228, 222, 222, 0), rgb(185, 27, 6));
-}
-
-/* .hologram > img {
-    animation: fade-in 0.7s ease-in backwards;
-    }
-    .hologram > img { 
-        animation: fade-out 0.7s ease-out backwards;
-        } */
-        
-        .hologram .false {
-    animation: fade-in 0.7s ease-in backwards;
-}
-
-.hologram .true { 
-    animation: fade-out 0.7s ease-out backwards;
-}
-
-@keyframes fade-in { 
-    0% {
-        opacity: 0%;
-    }
-    100% {
-        opacity: 100%;
-    }
-}
-
-@keyframes fade-out { 
-    0% { 
-      opacity: 100%;
-    }
-    100% {
-      opacity: 0%;
-    }
-  }
