+++ src/components/molecules/player-carousel/player-carousel.module.css
@@ -1,0 +1,49 @@
+.section {
+  height: 90%;
+  width: 100%;
+}
+
+.hologramSection {
+  margin-bottom: 20px;
+  z-index: -1;
+  opacity: 100;
+  height: 50%;
+  width: 100%;
+  background: linear-gradient(to top, rgba(228, 222, 222, 0), rgb(185, 27, 6));
+}
+
+.hologramSection > img {
+  height: 220%;
+  z-index: -1;
+
+}
+
+.fadeIn {
+  animation: fade-in 1s forwards;
+}
+
+.fadeOut {
+  animation-name: fade-out;
+  animation-duration: 4s;
+}
+
+
+.cardRow {
+  display: flex;
+  height: 50%;
+  overflow: scroll;
+  overflow-x: scroll;
+  scrollbar-width: none;
+  align-items: center;
+  scroll-snap-type: x mandatory;
+  z-index: 2;
+}
+
+@keyframes fade-in { 
+  from {
+      opacity: 0%;
+  }
+  to {
+      opacity: 100%;
+  }
+}
